Implement {here} macro support in _config.yml
It's important to have some macro in _config.yml which points to
config's directory, because otherwise we are unable to handle
correctly relative paths. For instance, we have the following
_config.yml:

    paths:
       content: ./
       output: _build/

and then we want to build a blog:

    $ holocron build --conf myblog/_config.yml

It's not obvious what user wants. Does he want to search for content
in current working directory or maybe in the folder with _config.yml?

Assumption that all relative paths in _config.yml are relative to
_config.yml is a poor assumption, and it seems the most rational
fix here is to introduce some helper macros so user will be able to
do what he really wants.

So the commit introduces the {here} macro which replaces with a path
to _config.yml's directory:

    paths:
       content: {here}/
       output:  {here}/_build/

Resolves #108

Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>