Add set_arguments method to Command interface
Each command has to be able to setup command-specific command line
arguments. This commit introduces a new "set_arguments" method in
the base Command class with the following signature:

    def set_arguments(self, parser):
        pass

If some command wants to add additional argument it has to override
this method and use provides 'parser' object to do this.

Signed-off-by: Igor Kalnitsky <igor@kalnitsky.org>