Add abc.Extension interface
Despite the fact Holocron doesn't care whether extension inherits base
classes (aka interfaces) until it quaks so, it's generally a good idea
to have interfaces anyway. Interfaces are useful in Python as a sort
of documentation, and they ensure you have implemented all needed
methods.

So this commit introduces an interface for Holocron extensions. It's
basically a callable that receives Holocron instance and do whatever it
wants (e.g. registers new converters and/or generators).