Implement API for theme extensions
The patch introduces API for registering new themes, that will be later
used for content rendering. So from now on it's allowed to create
third-party extensions which provide external themes.

As an example the ability to specify user theme is moved to extension.
It will be disabled by default in 0.4.0.