Rename "document-list.j2" -> "index.j2"
"document-list.j2" is not used by the application core but by builtin
extensions such as "index" or "tags". The purpose of this template is
to provide a standard way to generate an index page - a page that lists
posts matched some criteria.

So let's use more short and simple name "index" instead of quite a long
one "document-list".