Mark some templates as private with `_` prefix
From Holocron point of view there are two mandatory templates:

 * page.j2
 * post.j2

Yet, there's one more mandatory template if some built-in generators are
used – index.j2. Everything else is considered as implementation details
and should not be exposed. In order to mark them as private we name them
with leading underscore character.