Add basic facilities for processors pipeline
Processors pipeline is a new concept we want to go toward that allows
both to simplify Holocron core and to provide more flexibility in data
processing. Fron now one, everything goes through pipe of processors,
each receives a list of documents and returns a list of documents. It
does not matter what processors are going to do with it - remove
documents, change them or add new one - in the end what it matters that
output of one processor goes as input to another one.