Move document destination to document level
We used to calculate document's destination on fly when one requested
to build the document. This approach proved to be bad as rules are
hardcoded and can't be changed by third party extensions. Moving
destination into document model means from now on it can be changed if
default doesn't fit someone needs.