Move pretty URI logic out of core
Holocron is used to have a hardcoded logic for building pretty URIs,
i.e. URIs without '.html' suffix. As we are going toward processing
pipeline we don't want to see it to be a part of core. The logic in
core has to be as simple as possible, and any smart work needs to be
done via processors. So 'prettyuri' processor has been added to solve
this very specific issue - to build URIs for posts and pages without
'.html' suffix.