Replace Index generator w/ processor
The index page is a local HTML file that automatically loads when a web
browser starts and when the browser's 'home' button is pressed. The term
is also used to refer to the front page, web server directory index, or
main web page of a website.

Index processor generates an index page that contains links to a
documents (aka blog posts).