Check URLs and sitemap compatibility
The location of a Sitemap file determines the set of URLs that can be
included in that Sitemap. A Sitemap file located at

 http://example.com/catalog/sitemap.xml

can include any URLs starting with

 http://example.com/catalog/

but can not include URLs starting with

 http://example.com/images/

This commit introduces a check that raises an exception if a document
that can't be enlisted is filtered by Sitemap processor.