Use .updated instead of .updated_local in Sitemap
According to the Sitemap protocol, <lastmod> tag has to be in W3C
datetime format which is very similar to ISO-8601 but allows to omit
some parts. The protocol does not say a word about a timezone, so
let's don't be smart and use UTC. UTC is a good for understanding by
machines, and simplifies datetime management for developers. Let's
use timezones only where it matters.