Preserve known documents in Source processor
Source processor is used to find new documents, and add them to the
documents collection processed by processors pipeline. The thing is
it ought to preserve previously known documents, and only append new
ones, because there might be a case when you want to define few source
processors to search for documents in different places.