Do not use local datetime where not important
Every experienced developer knows: do not mess with datetimes, it's
trickier that it looks. One of the best practices here is to store
everything in UTC, doing conversion to local time (or any other tz)
on system edges.

Let's follow this advise in Holocron, and use UTC where we don't care
much about representation. One of such things here is Atom processor.
As for _local datetimes, let's depreacte them and remove in the future.