Add metadata processor
Metadata processor is a way to assign some key-value pairs on documents
so they can be later used by any other processors. One of such examples
may be 'template' attribute to set a template to be used by Jinja2. The
other one is, set 'group' and later use it in processors conditions.