Add pipeline processor
Pipeline processor is a good way to follow DRY and to collect all
processors that need to be applied for a certain (the same) set of
documents under one umbrella.

For instance, we want to:

 * set a custom template for all blog posts;
 * generate an Atom feed with all blog posts;
 * generate an index page with all blog posts;

In order to achieve this we need to specify 'when' condition for each
processor independently, which complicates readability and increase
maintenance cost since user need to keep all these conditions in-sync.

Pipeline processor, on the other hand, allows to specify 'when'
condition only once, and run processors pipeline for only matched set of
documents.