Screw backward compatibility
Well, I don't have much time to work on Holocron anyway, and since I'm
interested in changes rather in making everyone happy, let's do not
bother to be backward compatible this time. Besides, new Holocron
version will be so new and so different that there's no sense in
preserving one.