Add support for named pipelines
Named pipelines is a way to maintain more than one pipeline for the
site. There could be various use cases. For instance, a user may want
to have two building pipelines: one for production, and one for
development. Another example may be is that from now on, there's no need
to maintain commands as any command may be a pipeline.