Use JSON reference and schema for source processor
Holocron pipelines is a powerful mechanism to do whatever you wants,
however, it has one significant disadvantage - it's too verbose and
requires you to write a lot and even duplicate some things. In order
to overcome the issue, we came up with an idea of using JSON references,
so two or more processors can refer to the same source of truth defined
somewhere in global application context.

This is the first commit in a series designed to introduce new
facilities as well as set an example of how it may look like.

Along with JSON references, the commit also implements a schema
validation for input parameters.