Resolve existing pycodestyle warnings
Flake8 has been updated on Oct 24, 2018, bringing a new version of
pycodestyle as a dependency. It turns out that the latter has started to
complain if an invalid escape sequence is met in str/bytes with the
following warning:

 tests/ext/processors/test_restructuredtext.py:42:35: W605 invalid escape sequence '\s'
 tests/ext/processors/test_markdown.py:160:17: W605 invalid escape sequence '\s'

This is a preliminary step to force others to stop use Python the way it
wasn't intended because starting from Python 3.8 the lines that trigger
these warning will start treated by CPython as errors.

Fortunately, it's easy to get rid of these warnings simply by using
so-called raw string literals because such strings ignore escape
sequences letting them be propagated "As Is".