Fix error in one of test_source.py tests
One of test_source.py tests is a test bomb because it checks that
'Europe/Kyiv' timezone resolves into 'EEST' timezone. While this is true
during summer time, it's not true during winter time. In fact, it's not
true right now since it resolves into 'EET'. In order to fix the test
and make it more reliable, let's check that 'Europe/Kyiv' resolves into
either 'EET' or 'EEST'.