Use JSON reference and schema for index processor
Holocron pipelines is a powerful mechanism to do whatever you wants,
however, it has one significant disadvantage - it's too verbose and
requires you to write a lot and even duplicate some things. In order
to overcome the issue, we came up with an idea of using JSON references,
so two or more processors can refer to the same source of truth defined
somewhere in global application context.

This commit adds both JSON reference support and schema validation to
index processor.