Add fallback= option to @parameters decorator
This option is my attempt to achieve a good UX in Holocron by providing
consistent and DRY settings as well as being flexible enough to
overwrite each value on per-processor basis. Essentially, fallback= is a
way to say "hey, please try to extract a value from the metadata if it's
passed" and Holocron will do the rest. This is a safe attempt. If
nothing found under the specified key, a processor will use its default
value instead.