Drop dooku.ExtensionManager in favor of stevedore
Dooku is a set of libraries written by me in order to fill a gap in
Python standard library. Turns out, I never had a chance to make it
big and great. Now, we have boltons to fill that gap. Unfortunately
boltons does not support plugins, so we use stevedore instead which
is mature and stable.