Drop dependency on dooku.Conf in favor of dict
Dooku is a set of libraries written by me in order to fill a gap in
Python standard library. Turns out, I never had a chance to make it
big and great. Now, we have boltons to fill that gap. There's no real
need for dooku.Conf now when we support JSON pointers even in templates.