Add 'holocron.core.Item' class hierarchy
The hierarchy is designed to wrap a pipeline item around. So it's not a
dictionary anymore. It's an 'Item' instance instead. In most cases, no
one will notice that because the class hierarchy implements
MutableMapping interface, which means it can be used as a drop-in
replacement for 'dict'.

The main rationale behind this change is to introduce support for
automatically computed properties, like 'url' and 'absurl' that must be
computed based on 'destination' and 'baseurl' properties.

While this class wrapper is *not enforced*, it's recommended to use it
in processors. Who knows? Maybe one day we will make something smart
base on this.