Use 'core.Item()' instead of 'dict' in tests
Recently 'core.Item()' class was added to Holocron as a replacement for
plain dict items passed from one processor to another. The class mostly
behaves like a dict, however, it differs in a way that provides auto
computed items. While passing a dict through a pipeline is still legit,
we better use a special class instead to focus on our end goal.