Use 'black' to enforce code style
Black is the uncompromising Python code formatter. By using it, you
agree to cede control over minutiae of hand-formatting. In return, Black
gives you speed, determinism, and freedom from pycodestyle nagging about
formatting. You will save time and mental energy for more important
matters.

We do not introduce any pre commit hooks to ensure that the code is
formatted accordingly in this commit. Instead, we're adding a new PR
check to ensure that the code is Black formatted and it's up to
developers to call Black any way they prefer to.