Drop Python 3.5 support
Apparently due to using JSON schemas for validation some of our tests
flake. Annoyed by their behaviour I discovered that the flakes happen
because order of the keys in JSON schema is not guaranteed on CPython
before 3.6. This fact results in inconsistent validation order, leading
to inconsistent errors.

The workaround could be to rely on this internal knowledge and use
collections.OrderedDict initialized with a sequence of key-value pairs.
However, since Holocron is a pet project and is not used worldwide, I
think it would be better to drop Python 3.5 support (it's old anyways),
keep code readable and rely on keys order in the schema.