Pass theme options via jinja2's context
Processors must be as isolated as possible and hence do not rely on
things stored in metadata. If one wants to store something in metadata,
JSON reference must be used to tell processor to go and check metadata
instead.

That said, this commit moves theme options from metadata to jinja2's
processor context.