Move processors arguments under 'args' umbrella
Processors arguments are used to be defined top-level in the pipeline,
alongside with the processor name. This effectively means that every
processor definition in the pipeline has its own schema.

Well, I don't want to see that and I'd prefer to ensure that the schema
is consistent across definitions. One way to achieve this is to move
every processor argument under 'args' umbrella. This is exactly what
this commit is responsible for.