Add notion of processor wrappers
Processor wrappers are special kind of processors which receive another
processor as input argument and are fully responsible for invoking it.
In fact, they are just like any other processor except for syntax sugar
provided by Holocron that helps to invoke them easier and avoiding
nesting structures in pipeline definition.