Keep in .gitignore only relevant rules
I used to maintain a big .gitignore that are is likely to cover all
possible use cases. This was wrong, as it must only contain repository
specific rules. I've switched to per-user ignore configuration for
ignoring IDE related files a while ago, and this patch reduces the size
of .gitignore in this repository.